Refactor status panel module split
